split Booster.Builtin module into its parts
